fix: validation by pydantic and profile not found at creating posts

+++ src/yn/modules/posts/route.py
@@ -9,6 +9,7 @@ from yn.modules.posts.errors import (
 )
 from yn.modules.posts.schemas import PostCreate, PostRead, PostUpdate
 from yn.modules.posts.service import PostService
+from yn.modules.profiles.errors import ProfileNotFoundError
 from yn.modules.users.auth import get_current_user
 from yn.modules.users.dto import UserDTO
 
@@ -21,12 +22,15 @@ async def create_post(
     post_service: Annotated[PostService, Depends(get_post_service)],
     payload: PostCreate,
 ) -> PostRead:
+    if current_user.profile_id is None:
+        raise ProfileNotFoundError
+
     post = await post_service.create_post(
         profile_id=current_user.profile_id,
         title=payload.title,
         content=payload.content,
     )
-    return PostRead.model_validate(post, from_attributes=False)
+    return PostRead.model_validate(post, from_attributes=True)
 
 
 @router.put("/{post_id}")
@@ -36,6 +40,9 @@ async def update_post(
     post_service: Annotated[PostService, Depends(get_post_service)],
     payload: PostUpdate,
 ) -> dict[str, str]:
+    if current_user.profile_id is None:
+        raise ProfileNotFoundError
+
     if payload.title is None and payload.content is None:
         raise EmptyPostUpdateError
 
@@ -55,6 +62,9 @@ async def delete_post(
     current_user: Annotated[UserDTO, Depends(get_current_user)],
     post_service: Annotated[PostService, Depends(get_post_service)],
 ) -> dict[str, str]:
+    if current_user.profile_id is None:
+        raise ProfileNotFoundError
+
     await post_service.delete_post(post_id=post_id, profile_id=current_user.profile_id)
 
     return {"detail": "Post deleted successfully"}
@@ -66,7 +76,7 @@ async def search_posts(
     post_service: Annotated[PostService, Depends(get_post_service)],
 ) -> list[PostRead]:
     posts = await post_service.full_text_search_posts(query)
-    return [PostRead.model_validate(post, from_attributes=False) for post in posts]
+    return [PostRead.model_validate(post, from_attributes=True) for post in posts]
 
 
 @router.get("/")
@@ -74,4 +84,4 @@ async def get_posts(
     post_service: Annotated[PostService, Depends(get_post_service)],
 ) -> list[PostRead]:
     posts = await post_service.get_posts()
-    return [PostRead.model_validate(post, from_attributes=False) for post in posts]
+    return [PostRead.model_validate(post, from_attributes=True) for post in posts]
